Fix unused tiers indices
