+++ src/lib/public/board/data/terms.ts
@@ -1,0 +1,10 @@
+import { Term } from '../types';
+
+export const VISIBLE_TERMS = [
+  Term.Fall23,
+  Term.Spring23,
+  Term.Fall22,
+  Term.Spring22,
+  Term.Fall21,
+  Term.Spring21,
+];
